requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.2
python-dotenv==1.0.1
